+++ src/main/java/org/opensearch/knn/index/mapper/KNNVectorFieldMapperUtil.java
@@ -24,7 +24,7 @@ import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.KnnCircuitBreakerException;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.VectorDataType;
-import org.opensearch.knn.index.codec.util.KNNVectorSerializerFactory;
+import org.opensearch.knn.index.codec.util.KNNVectorAsCollectionOfFloatsSerializer;
 import org.opensearch.knn.index.engine.KNNEngine;
 import org.opensearch.knn.index.engine.KNNMethodContext;
 import org.opensearch.knn.index.engine.MethodComponentContext;
@@ -74,7 +74,7 @@ public class KNNVectorFieldMapperUtil {
      * @param vector vector to be added to stored field
      */
     public static StoredField createStoredFieldForFloatVector(String name, float[] vector) {
-        return new StoredField(name, KNNVectorSerializerFactory.getDefaultSerializer().floatToByteArray(vector));
+        return new StoredField(name, KNNVectorAsCollectionOfFloatsSerializer.INSTANCE.floatToByteArray(vector));
     }
 
     /**